Move transaction validity check to Format
